tests added and moved around
