+++ tests/test_output_files.py
@@ -167,266 +167,3 @@ def test_get_output_files_no_status_dir(test_dir):
     step = Step("mock-step", test_dir, step_settings)
     with pytest.raises(KeyError):
         out, err = step.get_output_files()
-
-
-@pytest.mark.parametrize(
-    "entity",
-    [pytest.param(ens, id="ensemble"), pytest.param(model, id="model")],
-)
-def test_symlink(test_dir, entity):
-    """Test symlinking historical output files"""
-    entity.path = test_dir
-    if entity.type == Ensemble:
-        for member in ens.models:
-            symlink_with_create_job_step(test_dir, member)
-    else:
-        symlink_with_create_job_step(test_dir, entity)
-
-
-def symlink_with_create_job_step(test_dir, entity):
-    """Function that helps cut down on repeated testing code"""
-    exp_dir = pathlib.Path(test_dir)
-    entity.path = test_dir
-    status_dir = exp_dir / CONFIG.telemetry_subdir / entity.type
-    step = controller._create_job_step(entity, status_dir)
-    controller.symlink_output_files(step, entity)
-    assert pathlib.Path(entity.path, f"{entity.name}.out").is_symlink()
-    assert pathlib.Path(entity.path, f"{entity.name}.err").is_symlink()
-    assert os.readlink(pathlib.Path(entity.path, f"{entity.name}.out")) == str(
-        status_dir / entity.name / (entity.name + ".out")
-    )
-    assert os.readlink(pathlib.Path(entity.path, f"{entity.name}.err")) == str(
-        status_dir / entity.name / (entity.name + ".err")
-    )
-
-
-@pytest.mark.parametrize(
-    "entity",
-    [
-        pytest.param(ens, id="ensemble"),
-        pytest.param(orc, id="orchestrator"),
-        pytest.param(anon_batch_model, id="model"),
-    ],
-)
-def test_batch_symlink(entity, test_dir):
-    """Test symlinking historical output files"""
-    exp_dir = pathlib.Path(test_dir)
-    entity.path = test_dir
-    status_dir = exp_dir / CONFIG.telemetry_subdir / entity.type
-    batch_step, substeps = slurm_controller._create_batch_job_step(entity, status_dir)
-    for step in substeps:
-        slurm_controller.symlink_output_files(step, entity)
-        assert pathlib.Path(entity.path, f"{entity.name}.out").is_symlink()
-        assert pathlib.Path(entity.path, f"{entity.name}.err").is_symlink()
-        assert os.readlink(pathlib.Path(entity.path, f"{entity.name}.out")) == str(
-            status_dir / entity.name / step.entity_name / (step.entity_name + ".out")
-        )
-        assert os.readlink(pathlib.Path(entity.path, f"{entity.name}.err")) == str(
-            status_dir / entity.name / step.entity_name / (step.entity_name + ".err")
-        )
-
-
-def test_symlink_error(test_dir):
-    """Ensure FileNotFoundError is thrown"""
-    bad_model = Model(
-        "bad_model",
-        params={},
-        path=pathlib.Path(test_dir, "badpath"),
-        run_settings=RunSettings("echo"),
-    )
-    telem_dir = pathlib.Path(test_dir, "bad_model_telemetry")
-    bad_step = controller._create_job_step(bad_model, telem_dir)
-    with pytest.raises(FileNotFoundError):
-        controller.symlink_output_files(bad_step, bad_model)
-
-
-def test_batch_model_and_ensemble(test_dir):
-    exp_name = "test-batch"
-    exp = Experiment(exp_name, launcher="slurm", exp_path=test_dir)
-
-    test_model = exp.create_model(
-        "test_model", path=test_dir, run_settings=batch_rs, batch_settings=bs
-    )
-    exp.generate(test_model)
-    exp.start(test_model, block=True)
-
-    assert pathlib.Path(test_model.path).exists()
-    _should_be_symlinked(pathlib.Path(test_model.path, f"{test_model.name}.out"), True)
-    _should_be_symlinked(pathlib.Path(test_model.path, f"{test_model.name}.err"), False)
-    _should_not_be_symlinked(pathlib.Path(test_model.path, f"{test_model.name}.sh"))
-
-    test_ensemble = exp.create_ensemble(
-        "test_ensemble", params={}, batch_settings=bs, run_settings=batch_rs, replicas=3
-    )
-    exp.generate(test_ensemble)
-    exp.start(test_ensemble, block=True)
-
-    assert pathlib.Path(test_ensemble.path).exists()
-    for i in range(len(test_ensemble.models)):
-        _should_be_symlinked(
-            pathlib.Path(
-                test_ensemble.path,
-                f"{test_ensemble.name}_{i}",
-                f"{test_ensemble.name}_{i}.out",
-            ),
-            True,
-        )
-        _should_be_symlinked(
-            pathlib.Path(
-                test_ensemble.path,
-                f"{test_ensemble.name}_{i}",
-                f"{test_ensemble.name}_{i}.err",
-            ),
-            False,
-        )
-
-    _should_not_be_symlinked(pathlib.Path(exp.exp_path, "smartsim_params.txt"))
-
-
-def test_failed_launch_symlinks(test_dir): ...
-
-
-def test_batch_ensemble_symlinks(test_dir):
-    exp_name = "test-batch-ensemble"
-    exp = Experiment(exp_name, launcher="slurm", exp_path=test_dir)
-    test_ensemble = exp.create_ensemble(
-        "test_ensemble", params={}, batch_settings=bs, run_settings=batch_rs, replicas=3
-    )
-    exp.generate(test_ensemble)
-    exp.start(test_ensemble, block=True)
-
-    for i in range(len(test_ensemble.models)):
-        _should_be_symlinked(
-            pathlib.Path(
-                test_ensemble.path,
-                f"{test_ensemble.name}_{i}",
-                f"{test_ensemble.name}_{i}.out",
-            ),
-            True,
-        )
-        _should_be_symlinked(
-            pathlib.Path(
-                test_ensemble.path,
-                f"{test_ensemble.name}_{i}",
-                f"{test_ensemble.name}_{i}.err",
-            ),
-            False,
-        )
-
-    _should_not_be_symlinked(pathlib.Path(exp.exp_path, "smartsim_params.txt"))
-
-
-def test_batch_model_symlinks(test_dir):
-    exp_name = "test-batch-model"
-    exp = Experiment(exp_name, launcher="slurm", exp_path=test_dir)
-
-    test_model = exp.create_model(
-        "test_model", path=test_dir, run_settings=batch_rs, batch_settings=bs
-    )
-    exp.generate(test_model)
-    exp.start(test_model, block=True)
-
-    assert pathlib.Path(test_model.path).exists()
-
-    _should_be_symlinked(pathlib.Path(test_model.path, f"{test_model.name}.out"), True)
-    _should_be_symlinked(pathlib.Path(test_model.path, f"{test_model.name}.err"), False)
-    _should_not_be_symlinked(pathlib.Path(test_model.path, f"{test_model.name}.sh"))
-
-
-def test_batch_orchestrator_symlinks(test_dir):
-    exp = Experiment("test-batch-orc", launcher="slurm", exp_path=test_dir)
-    port = 2424
-    db = exp.create_database(db_nodes=3, port=port, batch=True, single_cmd=False)
-    exp.generate(db)
-    exp.start(db, block=True)
-    exp.stop(db)
-
-    _should_be_symlinked(pathlib.Path(db.path, f"{db.name}.out"), False)
-    _should_be_symlinked(pathlib.Path(db.path, f"{db.name}.err"), False)
-
-    for i in range(db.db_nodes):
-        _should_be_symlinked(pathlib.Path(db.path, f"{db.name}_{i}.out"), False)
-        _should_be_symlinked(pathlib.Path(db.path, f"{db.name}_{i}.err"), False)
-        _should_not_be_symlinked(
-            pathlib.Path(db.path, f"nodes-orchestrator_{i}-{port}.conf")
-        )
-
-
-def test_non_batch_ensemble_symlinks(test_dir):
-    exp_name = "test-non-batch-ensemble"
-    rs = RunSettings("echo", ["spam", "eggs"])
-    exp = Experiment(exp_name, exp_path=test_dir)
-    test_ensemble = exp.create_ensemble(
-        "test_ensemble", params={}, run_settings=rs, replicas=3
-    )
-    exp.generate(test_ensemble)
-    exp.start(test_ensemble, block=True)
-
-    for i in range(len(test_ensemble.models)):
-        _should_be_symlinked(
-            pathlib.Path(
-                test_ensemble.path,
-                f"{test_ensemble.name}_{i}",
-                f"{test_ensemble.name}_{i}.out",
-            ),
-            True,
-        )
-        _should_be_symlinked(
-            pathlib.Path(
-                test_ensemble.path,
-                f"{test_ensemble.name}_{i}",
-                f"{test_ensemble.name}_{i}.err",
-            ),
-            False,
-        )
-
-    _should_not_be_symlinked(pathlib.Path(exp.exp_path, "smartsim_params.txt"))
-
-
-def test_non_batch_model_symlinks(test_dir):
-    exp_name = "test-non-batch-model"
-    exp = Experiment(exp_name, exp_path=test_dir)
-    rs = RunSettings("echo", ["spam", "eggs"])
-
-    test_model = exp.create_model("test_model", path=test_dir, run_settings=rs)
-    exp.generate(test_model)
-    exp.start(test_model, block=True)
-
-    assert pathlib.Path(test_model.path).exists()
-
-    _should_be_symlinked(pathlib.Path(test_model.path, f"{test_model.name}.out"), True)
-    _should_be_symlinked(pathlib.Path(test_model.path, f"{test_model.name}.err"), False)
-    _should_not_be_symlinked(pathlib.Path(exp.exp_path, "smartsim_params.txt"))
-
-
-def test_non_batch_orchestrator_symlinks(test_dir):
-    exp = Experiment("test-non-batch-orc", launcher="slurm", exp_path=test_dir)
-
-    db = exp.create_database(db_nodes=3, single_cmd=False)
-    exp.generate(db)
-    exp.start(db, block=True)
-    exp.stop(db)
-
-    for i in range(db.db_nodes):
-        _should_be_symlinked(pathlib.Path(db.path, f"{db.name}_{i}.out"), False)
-        _should_be_symlinked(pathlib.Path(db.path, f"{db.name}_{i}.err"), False)
-
-    _should_not_be_symlinked(pathlib.Path(exp.exp_path, "smartsim_params.txt"))
-
-
-def _should_not_be_symlinked(non_linked_path: pathlib.Path):
-    """Helper function for assertions about paths that should NOT be symlinked"""
-    assert non_linked_path.exists()
-    assert not non_linked_path.is_symlink()
-
-
-def _should_be_symlinked(linked_path: pathlib.Path, open_file: bool):
-    """Helper function for assertions about paths that SHOULD be symlinked"""
-    assert linked_path.exists()
-    assert linked_path.is_symlink()
-    # ensure the source file exists
-    assert pathlib.Path(os.readlink(linked_path)).exists()
-    if open_file:
-        with open(pathlib.Path(os.readlink(linked_path)), "r") as file:
-            log_contents = file.read()
-        assert "spam eggs" in log_contents
